refactor: address PR comments.. 1) remove EnvUtils module and move OjOptions directly under Fluent, 2) use strict=true in reformatted value and return the int parsing back to normal as that is a breaking change
