+++ lib/fluent/config/types.rb
@@ -111,7 +111,7 @@ module Fluent
     def self.symbol_value(val, opts = {}, name = nil)
       return nil if val.nil?
 
-      val[0] == ":" ? val[1..-1].to_sym : val.to_sym
+      val.delete_prefix(":").to_sym
     end
 
     SYMBOL_TYPE = Proc.new { |val, opts = {}, name = nil|
@@ -144,13 +144,7 @@ module Fluent
           raise ConfigError, "#{name}: #{e.message}"
         end
       else
-        i = val.to_i
-        # string will convert to 0, return nil if not actually "0"
-        if i == 0 && val != "0"
-          nil
-        else
-          i
-        end
+        val.to_i
       end
     }
 
@@ -196,7 +190,7 @@ module Fluent
         when :bool then Config.bool_value(value, opts, name)
         when :time then Config.time_value(value, opts, name)
         when :regexp then Config.regexp_value(value, opts, name)
-        when :symbol then SYMBOL_TYPE.call(value, opts, name)
+        when :symbol then Config.symbol_value(value, opts, name)
         else
           raise "unknown type in REFORMAT: #{type}"
         end
